# api/openapi.yaml
openapi: 3.0.3
info:
  title: RecurringTransactionsAPI
  description: API for analyzing bank CSV exports, detecting recurring payments, and providing annual overview dashboards with predictions.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/recurring-payments:
    get:
      operationId: getRecurringPayments
      summary: List all detected recurring payments
      tags: [RecurringPayments]
      responses:
        '200':
          description: List of recurring payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecurringPaymentDto'
    post:
      operationId: createRecurringPayment
      summary: Create a new recurring or grouped payment
      tags: [RecurringPayments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecurringPaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringPaymentDto'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/recurring-payments/simulate:
    post:
      operationId: simulateRules
      summary: Simulate rules against unlinked transactions
      tags: [RecurringPayments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulateRulesRequest'
      responses:
        '200':
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulateRulesResponse'
  /api/additional-rule-groups:
    get:
      operationId: getAdditionalRuleGroups
      summary: List Additional rule groups
      tags: [AdditionalRuleGroups]
      responses:
        '200':
          description: Additional rule groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdditionalRuleGroupDto'
    post:
      operationId: createAdditionalRuleGroup
      summary: Create an Additional rule group
      tags: [AdditionalRuleGroups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditionalRuleGroupRequest'
      responses:
        '201':
          description: Additional rule group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalRuleGroupMutationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate group name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/additional-rule-groups/{id}:
    get:
      operationId: getAdditionalRuleGroup
      summary: Get an Additional rule group
      tags: [AdditionalRuleGroups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Additional rule group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalRuleGroupDto'
        '404':
          description: Additional rule group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateAdditionalRuleGroup
      summary: Update an Additional rule group
      tags: [AdditionalRuleGroups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditionalRuleGroupRequest'
      responses:
        '200':
          description: Additional rule group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdditionalRuleGroupMutationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Additional rule group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate group name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteAdditionalRuleGroup
      summary: Delete an Additional rule group
      tags: [AdditionalRuleGroups]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Additional rule group deleted and matching recalculated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecalculationSummaryResponse'
        '404':
          description: Additional rule group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/recurring-payments/{id}:
    put:
      operationId: updateRecurringPayment
      summary: Update a recurring payment
      tags: [RecurringPayments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecurringPaymentUpdateRequest'
      responses:
        '200':
          description: Updated recurring payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringPaymentDto'
        '404':
          description: Recurring payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteRecurringPayment
      summary: Delete a recurring payment
      tags: [RecurringPayments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Payment deleted
        '404':
          description: Recurring payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/recurring-payments/{id}/transactions:
    get:
      operationId: getRecurringPaymentTransactions
      summary: Get transactions linked to a recurring payment
      tags: [RecurringPayments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of linked transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDto'
  /api/recurring-payments/{id}/history:
    get:
      operationId: getRecurringPaymentHistory
      summary: Get period amount history for a recurring payment
      tags: [RecurringPayments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of period history entries sorted by periodStart ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentPeriodHistoryEntry'
        '404':
          description: Recurring payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/recurring-payments/{recurringPaymentId}/rules:
    get:
      operationId: getRules
      summary: List rules for a recurring payment
      tags: [RecurringPaymentRules]
      parameters:
        - name: recurringPaymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleDto'
        '404':
          description: Recurring payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createRule
      summary: Create a new rule for a recurring payment
      tags: [RecurringPaymentRules]
      parameters:
        - name: recurringPaymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleRequest'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleDto'
        '400':
          description: Invalid rule configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recurring payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [status, message, timestamp]
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    RecalculationSummaryResponse:
      type: object
      required: [transactionsMarkedInterAccount, transactionLinksRemoved, recurringPaymentsDeleted, recurringPaymentsDetected]
      properties:
        transactionsMarkedInterAccount:
          type: integer
        transactionLinksRemoved:
          type: integer
        recurringPaymentsDeleted:
          type: integer
        recurringPaymentsDetected:
          type: integer
...
    TransactionDto:
      type: object
      required: [id, bookingDate, amount]
      properties:
        id:
          type: string
          format: uuid
        uploadId:
          type: string
          format: uuid
        bookingDate:
          type: string
          format: date
        account:
          $ref: '#/components/schemas/BankAccountDto'
        partnerName:
          type: string
        partnerIban:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        details:
          type: string
        isInterAccount:
          type: boolean
        linkedPaymentCount:
          type: integer
        linkedPaymentNames:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    BankAccountDto:
      type: object
      required: [id, iban]
      properties:
        id:
          type: string
          format: uuid
        iban:
          type: string
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Frequency:
      type: string
      enum: [MONTHLY, QUARTERLY, YEARLY]
    RuleType:
      type: string
      enum: [REGEX, JARO_WINKLER, AMOUNT]
    TargetField:
      type: string
      enum: [ACCOUNT, PARTNER_NAME, PARTNER_IBAN, DETAILS]
    RuleDto:
      type: object
      required: [id, ruleType]
      properties:
        id:
          type: string
          format: uuid
        ruleType:
          $ref: '#/components/schemas/RuleType'
        targetField:
          $ref: '#/components/schemas/TargetField'
        text:
          type: string
        strict:
          type: boolean
        threshold:
          type: number
          format: double
        amount:
          type: number
          format: double
        fluctuationRange:
          type: number
          format: double
    CreateRuleRequest:
      type: object
      required: [ruleType]
      properties:
        ruleType:
          $ref: '#/components/schemas/RuleType'
        targetField:
          $ref: '#/components/schemas/TargetField'
        text:
          type: string
        strict:
          type: boolean
        threshold:
          type: number
          format: double
        amount:
          type: number
          format: double
        fluctuationRange:
          type: number
          format: double
...
    PaymentType:
      type: string
      enum: [RECURRING, GROUPED]
    RecurringPaymentDto:
      type: object
      required: [id, name, averageAmount, frequency, isIncome, isActive, ruleCount, paymentType]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        categoryColor:
          type: string
        averageAmount:
          type: number
          format: double
        frequency:
          $ref: '#/components/schemas/Frequency'
        paymentType:
          $ref: '#/components/schemas/PaymentType'
        isIncome:
          type: boolean
        isActive:
          type: boolean
        ruleCount:
          type: integer
    RecurringPaymentUpdateRequest:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
        isActive:
          type: boolean
...
    CreateRecurringPaymentRequest:
      type: object
      required: [name, paymentType, frequency, rules]
      properties:
        name:
          type: string
        paymentType:
          $ref: '#/components/schemas/PaymentType'
        frequency:
          $ref: '#/components/schemas/Frequency'
        rules:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateRuleRequest'
    SimulateRulesRequest:
      type: object
      required: [rules]
      properties:
        draftType:
          $ref: '#/components/schemas/SimulationDraftType'
        currentAdditionalGroupId:
          type: string
          format: uuid
        rules:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateRuleRequest'
    SimulateRulesResponse:
      type: object
      required: [matchingTransactions, totalMatchCount, overlappingPayments]
      properties:
        matchingTransactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionDto'
        totalMatchCount:
          type: integer
        uniqueExclusionCount:
          type: integer
        omittedAdditionalMatchCount:
          type: integer
        omittedAdditionalMatches:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalGroupTransactionMatchDto'
        otherAdditionalGroupMatches:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalGroupTransactionMatchDto'
        overlappingPayments:
          type: array
          items:
            $ref: '#/components/schemas/OverlappingPaymentDto'
    SimulationDraftType:
      type: string
      enum: [RECURRING_PAYMENT, ADDITIONAL_GROUP]
    AdditionalRuleGroupRequest:
      type: object
      required: [name, rules]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        rules:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateRuleRequest'
    AdditionalRuleGroupDto:
      type: object
      required: [id, name, normalizedName, rules, excludedTransactionCount, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        normalizedName:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/RuleDto'
        excludedTransactionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdditionalRuleGroupMutationResponse:
      type: object
      required: [group]
      properties:
        group:
          $ref: '#/components/schemas/AdditionalRuleGroupDto'
        recalculationSummary:
          $ref: '#/components/schemas/RecalculationSummaryResponse'
    AdditionalGroupReferenceDto:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    AdditionalGroupTransactionMatchDto:
      type: object
      required: [transactionId, groups]
      properties:
        transactionId:
          type: string
          format: uuid
        groups:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalGroupReferenceDto'
    OverlappingPaymentDto:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    PaymentPeriodHistoryEntry:
      type: object
      required: [id, periodStart, periodEnd, amount]
      properties:
        id:
          type: string
          format: uuid
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        amount:
          type: number
          format: double